requirements.txt
streamlit>=1.30.0
pandas>=2.1.0
numpy>=1.25
mysql-connector-python==8.0.33
matplotlib==3.7.1
seaborn==0.12.2
google-generativeai==0.3.1
python-dotenv==1.0.0
joblib==1.2.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.2
opencv-python>=4.7.0
pillow>=9.5.0
scikit-learn>=1.2.2
tensorflow>=2.12.0
streamlit-chat>=0.0.2.2
transformers>=4.30.0
